+++ CHANGELOG.md
@@ -6,7 +6,9 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.19 (TBD)
 
-* TBD
+* Tile Scaling support for displaying missing tiles using nearby zoom levels
+* Zoom to tiles using the intersection between the Contents and Tile Matrix Set bounds
+* geopackage-android-map version updated to 2.0.2
 
 ## [1.18](https://github.com/ngageoint/geopackage-mapcache-android/releases/tag/1.18) (02-14-2018)